+++ pkg/provisioner/k8sprovisioner.go
@@ -93,6 +93,7 @@ func (p k8sProvisioner) newWorkerPod(args WorkerArgs) v1.Pod {
 	const (
 		repoVolumeName      = "repo"
 		repoVolumeMountPath = "/mnt/repo"
+		sshVolumeName       = "ssh"
 	)
 	labels := map[string]string{
 		"app":                          "wharf-cmd-worker",
@@ -110,6 +111,11 @@ func (p k8sProvisioner) newWorkerPod(args WorkerArgs) v1.Pod {
 			MountPath: repoVolumeMountPath,
 		},
 	}
+	gitVolumeMounts := append(volumeMounts, v1.VolumeMount{
+		Name:      sshVolumeName,
+		ReadOnly:  true,
+		MountPath: "/root/.ssh",
+	})
 
 	gitArgs := []string{"git", "clone", args.GitCloneURL, "--single-branch"}
 	if args.GitCloneBranch != "" {
@@ -184,7 +190,7 @@ func (p k8sProvisioner) newWorkerPod(args WorkerArgs) v1.Pod {
 					Image:           "bitnami/git:2-debian-10",
 					ImagePullPolicy: v1.PullIfNotPresent,
 					Command:         gitArgs,
-					VolumeMounts:    volumeMounts,
+					VolumeMounts:    gitVolumeMounts,
 				},
 			},
 			Containers: []v1.Container{
@@ -204,6 +210,16 @@ func (p k8sProvisioner) newWorkerPod(args WorkerArgs) v1.Pod {
 					Name: repoVolumeName,
 					VolumeSource: v1.VolumeSource{
 						EmptyDir: &v1.EmptyDirVolumeSource{},
+					},
+				},
+				{
+					Name: sshVolumeName,
+					VolumeSource: v1.VolumeSource{
+						Secret: &v1.SecretVolumeSource{
+							SecretName:  "wharf-cmd-worker-git-ssh",
+							DefaultMode: typ.Ref[int32](0600),
+							Optional:    typ.Ref(true),
+						},
 					},
 				},
 			},